add makefile

+++ 00/ex00/megaphone.cpp
@@ -12,7 +12,7 @@ int main(int argc, char *argv[])
     for (int i = 1; i < argc; i++)
     {
         std::string str = argv[i];
-        for (int j = 0; j < str.length(); j++)
+        for (int j = 0; j < (int)str.length(); j++)
           str[j] = toupper(str[j]);
         std::cout << str;
     }
